% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPMA.R
\docType{package}
\name{DPMA}
\alias{DPMA-package}
\alias{DPMA}
\title{DPMA Package}
\description{
Dynamic event prediction, using joint modeling of survival time and longitudinal variables, is extremely useful in personalized medicine. However, the estimation of joint models including many longitudinal markers is still a computational challenge because of the high number of random effects and parameters to be estimated. We use a model averaging strategy to combine predictions from several joint models for the event, including one longitudinal marker only   or pairwise longitudinal markers. The prediction is computed as the weighted mean of the predictions from the one-marker or two-marker  models, with the time-dependent weights estimated by minimizing the time-dependent Brier score. This method enables us to combine a large number of predictions issued from joint models to achieve a reliable and accurate individual prediction.
}
\references{
\enumerate{
\item
Hashemi, R., Baghfalaki, T., Philipps, V., and Jacqmin-Gadda, H. (2023). Dynamic prediction of an event using multiple longitudinal markers: a model averaging approach. \emph{Revised}.
}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, \email{taban.baghfalaki@u-bordeaux.fr}
}
\keyword{UDP}
