% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_DP.R
\name{summary_DP}
\alias{summary_DP}
\title{summary_DP function}
\usage{
summary_DP(UDPObject)
}
\arguments{
\item{UDPObject}{an object of class 'UDP' fitted by function UDP().}
}
\value{
\itemize{
\item DP: The values of dynamic prediction by utilizing model averaging for validation set (complement of training set) for each landmark time.
\item sd_Wald1: standard errors of the dynamic prediction based on Wald 1 approach.
\item sd_Wald2: standard errors of the dynamic prediction based on Wald 2 approach.
}
}
\description{
Compute the risk prediction base on  one-marker or two-marker model averaging.
}
\details{
It present the risk prediction base on  one-marker or two-marker model averaging.
}
\examples{
rm(list=ls())
data(pbc2_long)
data(pbc2_surv)


Modelmvglmer=list()
Modelmvglmer[[1]]=list(spiders ~  1+ year  + (1 + year | id))
Modelmvglmer[[2]]=list(albumin ~  1+ year  + (1 + year | id))
Modelmvglmer[[3]]=list(log(alkaline) ~  1+ year  + (1 + year | id))
Modelsurv=survival::Surv(years, status2) ~ 1

Results1=UDP(Modelsurv=Modelsurv, Modelmvglmer=Modelmvglmer, timeVar="year", nmark = 3, ncl = 1,
             families=c("binomial","gaussian","gaussian"), S = c(0,0.5,1), W = rep(0.5,3),
             index_training=c(1:250), datalong=pbc2_long,
             datasurv=pbc2_surv)


summary_UJMs(Results1,num_marker=2)
summary_DP(Results1)
\dontrun{
Results2=UDP(Modelsurv=Modelsurv, Modelmvglmer=Modelmvglmer, timeVar="year", nmark = 3, ncl = 5,
             families=c("binomial","gaussian","gaussian"), S = c(0,0.5,1), W = rep(0.5,3),
             index_training=c(1:250), datalong=pbc2_long,
             datasurv=pbc2_surv)


summary_UJMs(Results2,num_marker=2)
summary_DP(Results2)
}
}
\references{
R. Hashemi, T. Baghfalaki, V. Philipps, H. Jacqmin-Gadda. (2021). Dynamic prediction of an event using multiple longitudinal markers: a model averaging approach. \emph{Submitted}.
}
\author{
Taban Baghfalaki.
}
