% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_UJMs.R
\name{summary_UJMs}
\alias{summary_UJMs}
\title{summary_UJMs function}
\usage{
summary_UJMs(UDPObject, num_marker = 2)
}
\arguments{
\item{UDPObject}{an object of class 'UDP' fitted by function UDP().}

\item{num_marker}{Number of marker for presenting the summary statistics. If num_marker="All", the summaries of all markers will be given.}
}
\value{
\itemize{
\item Summary_UJM: summary statistics for each joint model (one-marker or two-marker).
}
}
\description{
Compute summary statistics for one-marker or two-marker joint models.
}
\details{
It present the summary statistics for one-marker or two-marker joint models.
}
\examples{
rm(list=ls())
data(pbc2_long)
data(pbc2_surv)


Modelmvglmer=list()
Modelmvglmer[[1]]=list(spiders ~  1+ year  + (1 + year | id))
Modelmvglmer[[2]]=list(albumin ~  1+ year  + (1 + year | id))
Modelmvglmer[[3]]=list(log(alkaline) ~  1+ year  + (1 + year | id))
Modelsurv=survival::Surv(years, status2) ~ 1

Results1=UDP(Modelsurv=Modelsurv, Modelmvglmer=Modelmvglmer, timeVar="year", nmark = 3, ncl = 1,
             families=c("binomial","gaussian","gaussian"), S = c(0,0.5,1), W = rep(0.5,3),
             index_training=c(1:250), datalong=pbc2_long,
             datasurv=pbc2_surv)


summary_UJMs(Results1,num_marker=2)
summary_DP(Results1)
\dontrun{
Results2=UDP(Modelsurv=Modelsurv, Modelmvglmer=Modelmvglmer, timeVar="year", nmark = 3, ncl = 5,
             families=c("binomial","gaussian","gaussian"), S = c(0,0.5,1), W = rep(0.5,3),
             index_training=c(1:250), datalong=pbc2_long,
             datasurv=pbc2_surv)


summary_UJMs(Results2,num_marker=2)
summary_DP(Results2)
}
}
\references{
R. Hashemi, T. Baghfalaki, V. Philipps, H. Jacqmin-Gadda. (2021). Dynamic prediction of an event using multiple longitudinal markers: a model averaging approach. \emph{Submitted}.
}
\author{
Taban Baghfalaki.
}
