% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UDP.R
\name{UDP}
\alias{UDP}
\title{UDP function}
\usage{
UDP(
  Modelsurv,
  Modelmvglmer,
  timeVar,
  nmark = nmark,
  ncl = 5,
  families,
  S = S,
  W = W,
  index_training,
  datalong,
  datasurv
)
}
\arguments{
\item{Modelsurv}{a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.}

\item{Modelmvglmer}{a list of R formulas representing the mixed models; these should be lme4-type formulas. See Examples.}

\item{timeVar}{a character string indicating the time variable in the multivariate mixed effects model.}

\item{nmark}{a scalar for the number of outcomes.}

\item{ncl}{the number of nodes to be forked for parallel socket cluster.}

\item{families}{a list of families objects correspond to each outcome.}

\item{S}{a vector for landmark times.}

\item{W}{a vector for prediction windows.}

\item{index_training}{index of individuals for the training set.}

\item{datalong}{a data.frame that contains all the variable to be used when fitting the multivariate mixed model.}

\item{datasurv}{a data.frame in which to interpret the variables named in the Modelsurv. The index of individual must be the same as those in datalong.}
}
\value{
\itemize{
\item RESULTS: The values of dynamic prediction by utilizing model averaging for validation set (complement of training set).
\item summaryJM_all: summary statistics for each joint model (one-marker or two-marker).
\item Comp_time: Computational time.
}
}
\description{
Compute dynamic prediction of an event using multiple longitudinal markers by applying a model averaging approach
}
\details{
It uses a model averaging strategy to combine predictions from several joint models for the event, including one longitudinal marker only or pairwise longitudinal markers. The prediction is computed as the weighted mean of the predictions from the one- or two-marker models, with the time-dependent weights estimated by minimizing the time-dependent Brier score.
}
\examples{
rm(list=ls())
data(pbc2_long)
data(pbc2_surv)


Modelmvglmer=list()
Modelmvglmer[[1]]=list(spiders ~  1+ year  + (1 + year | id))
Modelmvglmer[[2]]=list(albumin ~  1+ year  + (1 + year | id))
Modelmvglmer[[3]]=list(log(alkaline) ~  1+ year  + (1 + year | id))
Modelsurv=survival::Surv(years, status2) ~ 1

Results1=UDP(Modelsurv=Modelsurv, Modelmvglmer=Modelmvglmer, timeVar="year", nmark = 3, ncl = 1,
             families=c("binomial","gaussian","gaussian"), S = c(0,0.5,1), W = rep(0.5,3),
             index_training=c(1:250), datalong=pbc2_long,
             datasurv=pbc2_surv)


summary_UJMs(Results1,num_marker=2)
summary_DP(Results1)
\dontrun{
Results2=UDP(Modelsurv=Modelsurv, Modelmvglmer=Modelmvglmer, timeVar="year", nmark = 3, ncl = 5,
             families=c("binomial","gaussian","gaussian"), S = c(0,0.5,1), W = rep(0.5,3),
             index_training=c(1:250), datalong=pbc2_long,
             datasurv=pbc2_surv)


summary_UJMs(Results2,num_marker=2)
summary_DP(Results2)
}
}
\references{
R. Hashemi, T. Baghfalaki, V. Philipps, H. Jacqmin-Gadda. (2021). Dynamic prediction of an event using multiple longitudinal markers: a model averaging approach. \emph{Submitted}.
}
\author{
Taban Baghfalaki.
}
